refactor: extract server availability repository
