+++ app/src/main/java/org/ole/planet/myplanet/repository/ServerAvailabilityRepository.kt
@@ -1,0 +1,11 @@
+package org.ole.planet.myplanet.repository
+
+interface ServerAvailabilityRepository {
+    fun isPlanetAvailable(callback: PlanetAvailableListener?)
+
+    interface PlanetAvailableListener {
+        fun isAvailable()
+        fun notAvailable()
+    }
+}
+
